+++ DXLogQsoRecorder/StationConflictChoice.cs
@@ -1,0 +1,8 @@
+namespace DXLogQsoRecorder;
+
+public enum StationConflictChoice
+{
+    KeepCurrent,
+    SwitchToNew,
+    RecordAll
+}